// lightning-backend/src/api/bitcoin/bitcoin-api.interface.ts
export namespace IBitcoinApi {
// ...
  export interface Block {
    hash: string;                    //  (string) the block hash (same as provided)
    confirmations: number;           //  (numeric) The number of confirmations, or -1 if the block is not on the main chain
    size: number;                    //  (numeric) The block size
    strippedsize: number;            //  (numeric) The block size excluding witness data
    weight: number;                  //  (numeric) The block weight as defined in BIP 141
    height: number;                  //  (numeric) The block height or index
    version: number;                 //  (numeric) The block version
    versionHex: string;              //  (string) The block version formatted in hexadecimal
    merkleroot: string;              //  (string) The merkle root
    tx: Transaction[];
    time: number;                    //  (numeric) The block time expressed in UNIX epoch time
    mediantime: number;              //  (numeric) The median block time expressed in UNIX epoch time
    nonce: number;                   //  (numeric) The nonce
    bits: string;                    //  (string) The bits
    difficulty: number;              //  (numeric) The difficulty
    chainwork: string;               //  (string) Expected number of hashes required to produce the chain up to this block (in hex)
    nTx: number;                     //  (numeric) The number of transactions in the block
    previousblockhash: string;       //  (string) The hash of the previous block
    nextblockhash: string;           //  (string) The hash of the next block
  }

  export interface Transaction {
    in_active_chain: boolean;        //  (boolean) Whether specified block is in the active chain or not
    hex: string;                     //  (string) The serialized, hex-encoded data for 'txid'
    txid: string;                    //  (string) The transaction id (same as provided)
    hash: string;                    //  (string) The transaction hash (differs from txid for witness transactions)
    size: number;                    //  (numeric) The serialized transaction size
    vsize: number;                   //  (numeric) The virtual transaction size (differs from size for witness transactions)
    weight: number;                  //  (numeric) The transaction's weight (between vsize*4-3 and vsize*4)
    version: number;                 //  (numeric) The version
    locktime: number;                //  (numeric) The lock time
    vin: Vin[];
    vout: Vout[];
    blockhash: string;               //  (string) the block hash
    confirmations: number;           //  (numeric) The confirmations
    blocktime: number;               //  (numeric) The block time expressed in UNIX epoch time
    time: number;                    //  (numeric) Same as blocktime
  }

  export interface VerboseBlock extends Block {
    tx: VerboseTransaction[];        // The transactions in the format of the getrawtransaction RPC. Different from verbosity = 1 "tx" result
  }

  export interface VerboseTransaction extends Transaction {
    fee?: number;                   //  (numeric) The transaction fee in BTC, omitted if block undo data is not available
  }

  export interface Vin {
    txid?: string;                   //  (string) The transaction id
    vout?: number;                   //  (string)
    scriptSig?: {                    //  (json object) The script
      asm: string;                   //  (string) asm
      hex: string;                   //  (string) hex
    };
    sequence: number;                //  (numeric) The script sequence number
    txinwitness?: string[];          //  (string) hex-encoded witness data
    coinbase?: string;
    is_pegin?: boolean;              //  (boolean) Elements peg-in
  }

  export interface Vout {
    value: number;                   //  (numeric) The value in BTC
    n: number;                       //  (numeric) index
    asset?: string;                  //  (string) Elements asset id
    scriptPubKey: {                  //  (json object)
      asm: string;                   //  (string) the asm
      hex: string;                   //  (string) the hex
      reqSigs?: number;              //  (numeric) The required sigs
      type: string;                  //  (string) The type, eg 'pubkeyhash'
      address?: string;              //  (string) bitcoin address
      addresses?: string[];           //  (string) bitcoin addresses
      pegout_chain?: string;         //  (string) Elements peg-out chain
      pegout_addresses?: string[];   //  (string) Elements peg-out addresses
    };
  }
// ...
  export interface ChainTips {
    height: number;                  //  (numeric) height of the chain tip
    hash: string;                    //  (string) block hash of the tip
    branchlen: number;               //  (numeric) zero for main chain, otherwise length of branch connecting the tip to the main chain
    status: 'invalid' | 'headers-only' | 'valid-headers' | 'valid-fork' | 'active';
  }
// ...
}